requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.3.0
beautifulsoup4==4.12.3
boto3==1.35.98
botocore==1.35.98
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
Django==4.2.17
django-cors-headers==4.6.0
django-environ==0.12.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.55.3
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
lxml==5.3.0
openpyxl==3.1.5
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
pyphen==0.17.0
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
s3transfer==0.10.4
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xhtml2pdf==0.2.16
zopfli==0.2.3.post1